requirements.txt
selenium>=4.22.0
python-dotenv>=1.0.1
undetected-chromedriver>=3.5.5
urllib3<3
requests>=2.31.0
beautifulsoup4>=4.12.3
